[TL-82] Configured cut-through per port per queue
In the v3.3.0 release, cut-through is enabled only for queue0. Now it is
updated to enable cut through for queue 0 -7. The port number is
selected using the thousand for example 1000 is for lan1, 2000 for lan2
and so on.

+++ src/kernels/masternet-next/drivers/net/dsa/microchip/ksz_common.h
@@ -116,7 +116,7 @@ struct ksz_device {
 	u16 port_mask;
 
 	u8 tas_port;
-	u8 cut_through_enable;
+	u16 cut_through_enable;
 #if IS_ENABLED(CONFIG_NET_DSA_MICROCHIP_LAN937X_PTP)
 	struct ptp_clock_info ptp_caps;
 	struct ptp_clock *ptp_clock;
